# faster maximum, minimum (#28849)