--- a/test/reduce.jl
+++ b/test/reduce.jl
@@ -185,13 +185,21 @@ prod2(itr) = invoke(prod, Tuple{Any}, itr)
 @test minimum([4, 3, 5, 2]) == 2
 @test extrema([4, 3, 5, 2]) == (2, 5)
 
-@testset "maximum checks all elements" begin
+@testset "minimum/maximum checks all elements" begin
     N = 1025
     for i in 1:N
         arr = fill(0., N)
         truth = rand()
         arr[i] = truth
         @test maximum(arr) == truth
+
+        truth = -rand()
+        arr[i] = truth
+        @test minimum(arr) == truth
+
+        arr[i] = NaN
+        @test isnan(maximum(arr))
+        @test isnan(minimum(arr))
     end
 end