@chethega 2018-12-09
Looking at `@code_native`, it seems that vectorization fails?
```
julia> _fmax(x,y) = ifelse(x>y, x, y);
julia> function _fmaximum(arr)
              v=arr[1]
              @simd for i=1:length(arr)
              @inbounds v = _fmax(v,arr[i])
              end
              v
              end
julia> function _fmaximum2(arr)
              v1 = arr[1]
              v2 = arr[1]
              v3 = arr[1]
              v4=arr[1]
              isnan_ = isnan(v1) | isnan(v2)
              @simd for i=1:4:length(arr)
              @inbounds v1 = _fmax(v1,arr[i])
              @inbounds v2 = _fmax(v1,arr[i+1])
              @inbounds v3 = _fmax(v1,arr[i+2])
              @inbounds v4 = _fmax(v1,arr[i+3])
              isnan_ |= isnan(v1) | isnan(v2) | isnan(v3) | isnan(v4)
              end
              !isnan_ && return convert(typeof(v1),NaN)
              max(v1,v2,v3,v4)
              end

julia> function _fmaximum3(arr)
              v1 = arr[1]
              v2 = arr[1]
              v3 = arr[1]
              v4=arr[1]
              #isnan_ = isnan(v1) | isnan(v2)
              @simd for i=1:4:length(arr)
              @inbounds v1 = _fmax(v1,arr[i])
              @inbounds v2 = _fmax(v1,arr[i+1])
              @inbounds v3 = _fmax(v1,arr[i+2])
              @inbounds v4 = _fmax(v1,arr[i+3])
              #isnan_ |= isnan(v1) | isnan(v2) | isnan(v3) | isnan(v4)
              end
              #!isnan_ && return convert(typeof(v1),NaN)
              max(v1,v2,v3,v4)
              end


julia> using BenchmarkTools
julia> arr = randn(10^6);
julia> @btime maximum($arr);
  7.640 ms (0 allocations: 0 bytes)

julia> @btime _fmaximum($arr);
  1.341 ms (0 allocations: 0 bytes)

julia> @btime _fmaximum2($arr);
  1.021 ms (0 allocations: 0 bytes)

julia> @btime _fmaximum3($arr);
  537.057 μs (0 allocations: 0 bytes)

julia> arr = randn(Float32,10^6);

julia> @btime maximum($arr);
  7.233 ms (0 allocations: 0 bytes)

julia> @btime _fmaximum($arr);
  1.323 ms (0 allocations: 0 bytes)

julia> @btime _fmaximum2($arr);
  1.003 ms (0 allocations: 0 bytes)

julia> @btime _fmaximum3($arr);
  340.281 μs (0 allocations: 0 bytes)
```
Can we get properly vectorized (SIMD) code here? Does LLVM understand that min/max are associative?