@jw3126 2018-12-10
Oh nice @chethega. The timing of `_fmaximum3` looks like that of numpy. So I guess it uses SIMD? Should I update the PR with such a code? Or is this a compiler glitch, thats easy to fix?