@chethega 2018-12-10
I don't know. Your PR gives a nice 5x boost, but there is another big factor on the table. 

I guess some llvm expert needs to chime in on associativity of min/max algebra; did we write this in a bad way, are we missing an optimization pass or is this a problem of llvm? Or is this just stuck [upstream](https://reviews.llvm.org/rL345220) and will solve itself soon? 

If the latter, then I'd say merge and keep an issue open to revisit this later, once the fundamentals in llvm are ready. Nobody got grey hair waiting for the slow `maximum` yet, so people can enjoy the 5x speedup until they get the next 4x; salami tactics yay. 

Second point: I don't understand why your implementation is correct with `NaN`. 
```
julia> fm(x,y)=ifelse(x > y, x, y)
julia> A=[1.0, NaN, 2.0];
julia> fm(fm(A[1],A[2]), A[3])
2.0
```
Does your implementation handle that correctly? Regardless, could you add that test-case, just for my piece of mind? I think this is an example that any future PR can plausibly get wrong, which makes it a good test (even better test: `[1.0, 2.0, NaN, 4.0, 5.0]`).