@jw3126 2018-12-10
Yes the implentation should be `NaN` correcect, there are also tests. The trick is to ensure, that all `NaN` are either the second argument or checked explicitly.
```julia
julia> fm(x,y)=ifelse(x > y, x, y)
fm (generic function with 1 method)

julia> A=[1.0, NaN, 2.0];

julia> fm(fm(A[1],A[2]), A[3])
2.0

julia> fm(A[3], fm(A[1],A[2]))
NaN
```