@chethega 2018-12-10
My problem is that I don't see how you ensure that. If you fold a `NaN` into a number, you get a `NaN`; good. But if you fold a number into a `NaN`, then you get the number, and since you only check every `chunk_len`, one of the `NaN`s could escape.

```
julia> @eval Base _fast(::typeof(max), x, y) = ifelse(x > y, x, y)
julia> @eval Base _fast(::typeof(min), x, y) = ifelse(x < y, x, y)

julia> @eval Base function mapreduce_impl(f::ft, op::Union{typeof(max), typeof(min)},A::AbstractArray, first::Int, last::Int) where ft
        a1 = @inbounds A[first]
        v = mapreduce_first(f, op, a1)
        chunk_len = 256
        stop = first
        while last - stop > 0
        v == v || return v
        start = stop + 1
        stop = min(stop + chunk_len, last)
        for i in start:stop
        @inbounds ai = A[i]
        v = _fast(op, v, f(ai))
        end
        end
        v
        end

julia> A=rand(5); A[3]=NaN; maximum(A)
NaN
julia> A=rand(1000); A[400]=NaN; maximum(A)
0.9991780990981842
```