@chethega 2018-12-10
That looks correct, but gives almost a 2x slowdown compared to the original PR :(

Somewhat faster variant for floats on my machine is
```
julia> function _maximum(a::AbstractArray{T}) where T
       v0=typemin(T)
       i1 = false
       rs = mapreduce(_lift, _red, a; init=(v0,i1))
       rs[2] && return T(NaN)
       rs[1]
       end
julia> fmax(x,y)=ifelse(x > y, x, y)
julia> _lift(x)=(x, isnan(x));
julia> _red(a,b)=( (v1,i1) = a; (v2,i2)=b; (fmax(v1,v2),i1|i2) )
julia> @btime _maximum($arr)
  1.621 ms (0 allocations: 0 bytes)
```
I think this is because (1) the above `fmax` is a hardware instruction, and (2) we are latency constrained. That is, apart from the maybe upstream issue of using SIMD / vector units, we have all these nice execution ports idling around, because the code is written such that we have one giant dependency chain. The above variant breaks at least the nan-check out of that dependency chain.