@jw3126 2018-12-10
That is weird, on my machine, it is the same speed as original PR.