@chethega 2018-12-11
Nice, no more objections from me. We might want a `@fastmath` as well, for skipping nan-handling and signed zeros. I think passing an `Ordering` is a cool idea but should be a separate feature PR.

Can we get a nanosoldier run? Just in case of surprises?