@nanosoldier 2018-12-12
[Your benchmark job](https://github.com/JuliaLang/julia/pull/30320#issuecomment-446386177) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/8f8a11eac303ea012023e97921813acda0546400/39cf3c2_vs_caa0273/report.md). cc @ararslan