@KristofferC 2018-12-18
Right now we are only backporting bugfixes to 1.1.0. This can go in 1.1.1.