@ViralBShah 2018-12-18
Reverting due to the reported breakage. Probably should also be added to tests.